src/main/java/com/siukatech/poc/react/backend/parent/business/service/UserService.java
package com.siukatech.poc.react.backend.parent.business.service;

import com.siukatech.poc.react.backend.parent.data.entity.UserEntity;
import com.siukatech.poc.react.backend.parent.data.repository.UserRepository;
import com.siukatech.poc.react.backend.parent.business.dto.UserDto;
import jakarta.persistence.EntityNotFoundException;
import org.modelmapper.ModelMapper;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.stereotype.Service;

@Service
public class UserService {

    private final Logger logger = LoggerFactory.getLogger(this.getClass());
    private final ModelMapper modelMapper;
    private final UserRepository userRepository;

    private UserService(ModelMapper modelMapper, UserRepository userRepository) {
        this.modelMapper = modelMapper;
        this.userRepository = userRepository;
    }

    public UserDto findByUserId(String targetUserId) {
        UserEntity userEntity = this.userRepository.findByUserId(targetUserId)
                .orElseThrow(() -> new EntityNotFoundException("No such user [%s]".formatted(targetUserId)));
        logger.debug("findByUserId - modelMapper: [" + this.modelMapper + "]");
        UserDto userDto = this.modelMapper.map(userEntity, UserDto.class);
        return userDto;
    }

}
